requirements.txt
uvicorn==0.23.0
fastapi==0.103.2
pydantic==2.4.2
typing-extensions==4.8.0
experta==1.9.4
frozendict==2.3.8
yfinance==0.2.31
pgmpy==0.1.22
sqlalchemy==2.0.21
bcrypt==4.0.1
python-dotenv==1.0.0
passlib==1.7.6
jose==3.3.0
psycopg2==2.9.8
python-jose==3.3.0
pydantic[email]==2.4.2
python-multipart==0.0.6